+++ app/src/main/java/sancho/gnarlymusicplayer/MediaPlaybackService.kt
@@ -23,6 +23,7 @@ import androidx.core.app.NotificationManagerCompat
 import androidx.media.AudioAttributesCompat
 import androidx.media.AudioFocusRequestCompat
 import androidx.media.AudioManagerCompat
+import sancho.gnarlymusicplayer.models.Track
 import java.io.IOException
 
 class MediaPlaybackService : Service()
