+++ app/src/main/java/sancho/gnarlymusicplayer/models/Track.kt
@@ -1,3 +1,3 @@
-package sancho.gnarlymusicplayer
+package sancho.gnarlymusicplayer.models
 
 class Track(var path: String, var name: String)
